backend/requirements.txt
fastapi[standard]==0.115.0
sqlmodel==0.0.18
PyJWT[crypto]==2.9.0
python-dotenv==1.0.1